Proof of concept
Fixes #1

Need to verify the technical feasibility of the project. Especially around maintained JS libraries that will allow this all to happen on the client side.

No major issues so far, not super happy with the jimp libraries, but
ultimately this will be in a worker and the image manipulation can be
switched to a different library if needed (image-js?).