Fix glasses cache issue when resizing
If size of the glasses is changed, the cache would not take that into
account and it would lead to unexpected results.
Now, size of the glasses is part of the key.
Might need to think what would be the best approach here: size is
cached, but flip not. Maybe to keep it simple, at some point I'll switch
to just caching the original images and then resize, flip and possibly
do other things at every render.