Clip glasses when they are outside parent
Follow-up to ffb75004f850c33aba7235d4815a7e9800ea8ede
Fixes #24 (along with the above commit)

Used `clip-path` to make sure that while the glasses can move slightly
outside the parent, the part that is outside the parent, will not be
visible. This approach is a bit more complicated than simply having some
padding and z-index. But it gives much more flexibility - the amount of
offset the glasses can move outside the parent depends on their
size/scale. With the padding approach, that would not be possible. So
it's a lucky coincidence that the more complicated approach solved that
elegantly.