Revert "Revert "Allow changing of colors for Classic glasses""
This reverts commit 4372740e1fee2df86ba0afc4af5a5c13ae12557e.

Which brings back the glasses color changing functionality.
The root issue was that if the color was not changed from the default
black, the SVG image would be used to generate the output image. But
gifwrap does not support that. To potentially avoid the conversion by
default, let's use a pre-generated PNG image for the default, black
glasses. And then when the color change is requested, generate on the
fly the data URL from the SVG.

Related to https://github.com/klimeryk/dealwithit/issues/33